Integration tests cont.

+++ src/StoreManager.Application/Commands/Chain/Handlers/DeleteChainCommandHandler.cs
@@ -15,6 +15,10 @@ public class DeleteChainCommandHandler : ICommandHandler<DeleteChainCommand>
 
     public async Task<Result> Handle(DeleteChainCommand command, CancellationToken cancellationToken = default)
     {
+        if (_chainRepository.GetCountofStoresByChainAsync(command.Id).Result > 0)
+        {
+            return Result.Fail(Errors.ChainErrors.ChainHasStores());
+        }
         await _chainRepository.DeleteAsync(command.Id, cancellationToken);
 
         return Result.Ok();
